Allow externally provided epsilon factors thus enabling tNCS correction.

+++ cctbx/xray/boost_python/targets.cpp
@@ -163,7 +163,7 @@ namespace {
           af::const_ref<double> const&,
           af::const_ref<double> const&,
           double,
-          af::const_ref<int> const&,
+          af::const_ref<double> const&,
           af::const_ref<bool> const&,
           bool>((
             arg("f_obs"),
